refactor: replace FirebaseCheckDatabaseService with FirestoreCheckDatabaseService for user-specific checks
